Another attempt to fix #35

+++ YogaSMC/YogaVPC/IdeaVPC.hpp
@@ -279,7 +279,7 @@ private:
      *
      *  @return true if success
      */
-    bool read_ec_data(UInt32 cmd, UInt32 *result, UInt8 *retries);
+    bool read_ec_data(UInt32 cmd, UInt32 *result, UInt32 *retries);
 
     /**
      *  Write EC data
@@ -290,7 +290,7 @@ private:
      *
      *  @return true if success
      */
-    bool write_ec_data(UInt32 cmd, UInt32 value, UInt8 *retries);
+    bool write_ec_data(UInt32 cmd, UInt32 value, UInt32 *retries);
 
     /**
      *  Write EC data
